move ErrorHelper to PackageDialog
